+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Added `colors` property to M2 metadata containing color/alpha timelines for models.
 - Added `textureWeights` property to M2 metadata containing global texture transparency timelines for models.
 - Added `transparencyLookup` property to M2 metadata for texture unit mapping to transparency entries.
+- Added `textureTransforms` property to M2 metadata defining texture animation timelines.
+- Added `textureTransformsLookup` property to M2 metadata for texture transform lookups.
 
 0.1.35 (30-08-2021)
 - Fixed issue that caused invalid OBJ files to be produced for M2 collision geometry.